Switch handlers again to be on its own separate file in order to correctly set commons and rpc client reference
